Fix bug with AnsiConsole.Markup
Interpolated strings are passed to AnsiConsole.Markup to display matches and errors. This includes the text of a matching line. This text may contain "[" characters, which `AnsiConsole` tries to treat as a style. This is at best going to produce text styled in an unpredictable and unintended manner, and more likely will cause an exception to be thrown when the "style" isn't found.

Instead we should call `MarkupInterpolated` which will handle escaping these kinds of characters so they _aren't_ treated as style information.